+++ R/plotPCA-methods.R
@@ -1,6 +1,6 @@
 #' @name plotPCA
 #' @inherit AcidPlots::plotPCA
-#' @note Updated 2021-03-15.
+#' @note Updated 2022-01-21.
 #'
 #' @details Passes to `SummarizedExperiment` defined in AcidPlots package.
 #'
@@ -13,61 +13,22 @@ NULL
 
 
 
-## Updated 2019-09-10.
+## Updated 2021-01-21.
 `plotPCA,DESeqAnalysis` <-  # nolint
     function(object, ...) {
         validObject(object)
         dt <- as(object, "DESeqTransform")
-        plotPCA(object = dt, ...)
-    }
-
-
-
-## Keep this here so we don't allow inheritance of SE method.
-## Updated 2021-03-15.
-`plotPCA,DESeqDataSet` <-  # nolint
-    function(object, ...) {
-        abort(sprintf(
-            "Use {.cls %s} object instead of {.cls %s}.",
-            "DESeqTransform", "DESeqDataSet"
-        ))
-    }
-
-
-
-## Updated 2019-07-23.
-`plotPCA,DESeqTransform` <-  # nolint
-    function(object, ...) {
-        rse <- as(object, "RangedSummarizedExperiment")
+        rse <- as(dt, "RangedSummarizedExperiment")
         plotPCA(rse, ...)
     }
 
 
 
-#' @describeIn plotPCA Extracts `DESeqTransform` and passes to corresponding
-#'   method.
+#' @describeIn plotPCA Extracts `DESeqTransform`, converts to
+#'   `RangedSummarizedExperiment`, and passes to method defined in AcidPlots.
 #' @export
 setMethod(
     f = "plotPCA",
     signature = signature(object = "DESeqAnalysis"),
     definition = `plotPCA,DESeqAnalysis`
-)
-
-#' @describeIn plotPCA Method intentionally errors. Use `DESeqAnalysis` or
-#'   `DESeqTransform` methods instead.
-#' @export
-setMethod(
-    f = "plotPCA",
-    signature = signature(object = "DESeqDataSet"),
-    definition = `plotPCA,DESeqDataSet`
-)
-
-#' @describeIn plotPCA Passes to `SummarizedExperiment` method defined in
-#'   AcidPlots package. Uses values defined in
-#'   [`assay()`][SummarizedExperiment::assay].
-#' @export
-setMethod(
-    f = "plotPCA",
-    signature = signature(object = "DESeqTransform"),
-    definition = `plotPCA,DESeqTransform`
 )